- Implementing the StateFlow to handle ProductFormScreen UI changes with the source of data (dao)
