+++ app/src/main/java/com/alura/aluvery/ui/activities/ProductFormActivity.kt
@@ -3,21 +3,23 @@ package com.alura.aluvery.ui.activities
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.viewModels
 import androidx.compose.material.Surface
-import com.alura.aluvery.dao.ProductDao
 import com.alura.aluvery.ui.screens.ProductFormScreen
 import com.alura.aluvery.ui.theme.AluveryTheme
+import com.alura.aluvery.ui.viewmodels.ProductFormScreenViewModel
 
 class ProductFormActivity : ComponentActivity() {
-    private val dao = ProductDao()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             AluveryTheme {
                 Surface {
+                    val viewModel: ProductFormScreenViewModel by viewModels()
                     ProductFormScreen(
-                        onSaveClick = { product ->
-                            dao.save(product)
+                        viewModel = viewModel,
+                        onSaveClick = {
                             finish()
                         }
                     )
