+++ app/src/main/java/com/alura/aluvery/ui/viewmodels/ProductFormScreenViewModel.kt
@@ -1,0 +1,95 @@
+package com.alura.aluvery.ui.viewmodels
+
+import androidx.lifecycle.ViewModel
+import com.alura.aluvery.dao.ProductDao
+import com.alura.aluvery.model.Product
+import com.alura.aluvery.ui.states.ProductFormUiState
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
+import java.math.BigDecimal
+import java.text.DecimalFormat
+
+class ProductFormScreenViewModel : ViewModel() {
+
+    private val dao = ProductDao()
+
+    private val _uiState: MutableStateFlow<ProductFormUiState> = MutableStateFlow(
+        ProductFormUiState()
+    )
+
+    val uiState get() = _uiState.asStateFlow()
+
+    private val formatter = DecimalFormat("#.##")
+
+    init {
+        _uiState.update { currentState ->
+            currentState.copy(
+                onUrlChange = {
+                    _uiState.value = _uiState.value.copy(
+                        url = it,
+                    )
+                },
+                onNameChange = {
+                    _uiState.value = _uiState.value.copy(
+                        name = it,
+                    )
+                },
+                onPriceChange = {
+
+                    val isPriceError = try {
+                        BigDecimal(it)
+                        false
+                    } catch (e: IllegalArgumentException) {
+                        it.isNotEmpty()
+                    }
+
+                    val price = try {
+                        formatter.format(BigDecimal(it))
+                    } catch (e: IllegalArgumentException) {
+                        if (it.isEmpty()) {
+                            it
+                        } else {
+                            null
+                        }
+                    }
+
+                    price?.let {
+                        _uiState.value = _uiState.value.copy(
+                            price = price,
+                        )
+                    }
+
+                    _uiState.value = _uiState.value.copy(
+                        isPriceError = isPriceError,
+                    )
+
+                },
+                onDescriptionChange = {
+                    _uiState.value = _uiState.value.copy(
+                        description = it,
+                    )
+                }
+
+            )
+        }
+
+    }
+
+    fun save() {
+        uiState.value.run {
+            val convertedPrice = try {
+                BigDecimal(price)
+            } catch (e: NumberFormatException) {
+                BigDecimal.ZERO
+            }
+            val product = Product(
+                name = name,
+                image = url,
+                price = convertedPrice,
+                description = description
+            )
+            dao.save(product)
+        }
+    }
+}
